requirements.txt
appdirs==1.4.4
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.0
fonttools==4.59.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mss==10.0.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycryptodome==3.21.0
pynput==1.7.7
pyparsing==3.2.3
pyshark==0.6
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2025.1
requests==2.32.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.16.0
termcolor==2.5.0
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.2.3
Werkzeug==3.1.3
PyYAML
